changed after rewiew

+++ HW1/Yhrechykhin/HW1_1.py
@@ -1,10 +1,9 @@
-# 1
 a = 10
 b = 3
 
-print(int(a + b))
-print(int(a - b))
-print(int(a * b))
-print(int(a / b))
-print(int(a ** b))
+print(a + b)
+print(a - b)
+print(a * b)
+print(a / b)
+print(a ** b)
